code refactored for default node
